+++ flag.go
@@ -4,8 +4,17 @@ import (
 	"flag"
 	"regexp"
 
+	"github.com/ozontech/testo/internal/env"
+	"github.com/ozontech/testo/internal/parse"
+
 	// so that cache flags are always available.
 	_ "github.com/ozontech/testo/testocache"
+)
+
+var flagStrict = flag.Bool(
+	"testo.strict",
+	parse.Bool(env.TestoStrict.Get()),
+	"turn warnings into errors",
 )
 
 var flagMethod = flagRegexp{Regexp: regexp.MustCompile("")}