+++ testocache/cache.go
@@ -25,17 +25,20 @@ import (
 	"slices"
 	"strconv"
 	"sync"
+
+	"github.com/ozontech/testo/internal/env"
+	"github.com/ozontech/testo/internal/parse"
 )
 
 var (
 	flagDir = flag.String(
 		"cache.dir",
-		cmp.Or(os.Getenv("TESTO_CACHE_DIR"), ".testo_cache"),
+		cmp.Or(env.TestoCacheDir.Get(), ".testo_cache"),
 		"directory where the testo cache is stored",
 	)
 	flagDisable = flag.Bool(
 		"cache.disable",
-		parseBool(os.Getenv("TESTO_CACHE_DISABLE")),
+		parse.Bool(env.TestoCacheDisable.Get()),
 		"disable caching in testo",
 	)
 )
@@ -280,12 +283,6 @@ func cacheDir() (string, error) {
 	}
 
 	return dir, nil
-}
-
-func parseBool(s string) bool {
-	b, _ := strconv.ParseBool(s)
-
-	return b
 }
 
 func validate(key string) error {